Added character actions and logs for the same.
Added Say, Action, Omit Say, Omit Action, Test Say, and Test Action to derivatives. Says are in quotes and unbolded, while actions are bolded and given as a sentence (or sentences) starting with the character name. Omits and Tests are put into the OOC, and Tests do not display for other players. The log files have been altered to include all these features accordingly, and now include buttons to only display IC posts or OOC posts as desired.