Better face icons, cleanups, Players Online, and Whispers.
Altered canvas CSS for somewhat better face icon output, and altered the way face icons were handled to be more like CSS sprites, drastically reducing the number of http calls. Various cleanups and minor costless optimizations were performed. A very basic Players Online list is present, with names that can be right clicked for a Whisper function. Some backend work on permissions and forum infrastructure has been done, but little to none of it is functional as of yet.