Sanitized HTML support, various quality of life updates.
Basic text HTML can now be used in posts, and characters now have a Custom HTML option to replace their name with a formatted version in posts. Rules and Message of the Day now display on login, and can be configured by admins from the Admin Commands list or displayed again by players from the Player Commands tab. An OOC message is now sent to a user when their permissions level changes, and reconnects now automatically attempt to log back in. On a more minor note, Set OOC Color now immediately changes the color upon selection, and dice now output a sum total as well as the individual rolls.