Character profiles.
Performed some minor formatting and checking fixes, made the MOTD and Rules appear in the textarea when edited by an admin. Added a default character profile that can be edited in the same way as the MOTD and Rules, and added a Set Profile command to characters to modify their linked profile. Character face icons can be clicked to view their profile if it exists, and the Player Commands tab has an Open Character Database function to navigate all existent profiles.