Room functionality.
Added Join Room and Leave Room functions to characters. Characters in rooms send their IC posts to the room in question, and they are not logged. The Set Room function in the Player Commands tab allows a player to join a room without a specific character and determines which room, if any, their Narrate posts go to. Show Rooms displays all of the rooms that the user or their characters are in by name.