Started cleaning up modal appearance, added drag and drop.
Name tags have been added for posts into rooms, and titles are now displayed at the top of modals to show the function and name of the character, group, or derivative they are acting on, if applicable. Characters, groups, and derivatives can also be dragged and dropped to reorder them or, in the case of characters, move them into and out of groups.