World Info, login, and cleanup
Added a World Info page that appears on login, which now fills the screen to prevent misleading access to the UI prior to logging in. It can be edited from Admin Commands and viewed from Player Commands as well. Show MOTD and Show Rules no longer fail when blank, character and group dragging has been made somewhat easier, customHTML can be set blank to default back to the normal name, modals can no longer be dragged outside the space of the window, and character creation modals will no longer horizontally change size if a large number of derivatives are input.