Various small fixes.
Fixed oversights in post formatting in the logs and for actions. In a slightly more important case, fixed the wrong derivative being used for deletion, editing, or posts if the derivative list for the given character is rearranged. Also prevented an infinite loop case with dice creation, and added experimental protections against abuse of the style attribute in posts and names.