Added idle timer and notification settings, some fixes.
An idle timer has been added that automatically lists users as AFK if they do nothing in the client tab for long enough. Also added an option in the Player Commands tab to set which users will notify you with their posts, changing the page title and highlighting the posts until they have been read. An issue with the playerlist trying to update permissions while the client is not logged in and a minor formatting issue with editing Say posts have both been fixed.