Post Deletion and log dispay changes.
Added a Delete Post option to remove posts from the IC window. They are marked as deleted in the logs, but still appear.
Adjusted how edits show up in the logs. They display instead of the original post, but can be clicked to switch to the original post. Only the latest edit displays.