Added two QoL options.
There is now a Remember me option on login to autofill username and password on subsequent logins. This requires cookies to be permitted.
A Show Default Profile command has been added to the Player Commands, which opens a modal containing the default character profile to be copied for personal use without the need to create a new character.