Removed functional dependencies on IPs.
The database now has its own login, which is only required for editing. An
active login is no longer necessary for viewing the database, and as a
result, several dependencies on using IP for identification in the server
have been removed as unnecessary. Multiple connections on the same IP
should now be possible, and no further errors should be caused by changes
in IP.