Admin logs, admin profile modification.
Added admin logs, which log any use of admin commands on the server except for the viewing functions such as List Bans. An option has also been added to acquire a list of usernames and characters in Admin Commands. The character profiles can be modified or deleted from here at admin discretion, which is also logged. This is a preliminary measure before any formal approval functionality is added.