Upgraded Notify options and Scroll mode.
Right clicking a character post or narration now allows a Notify for Player option, and if a character or player is already on the notification list, this and Notify for Character are replaced by a Stop Notifying option. Also, Scroll mode on Active Character modals now includes two buttons to adjust the number of rows of face icons from the default of two.