Cleaned up the database.
Editing modals now resize in a manner identical to the main UI, removing some issues with textarea resizing. Logins occur automatically if cookies are present, and logging in with the modal or otherwise no longer results in popup alerts. An area directly above the database directories displays the connectivity state instead.