Better checking for automatic line breaks.
The helper regex for automatic line breaking has been improved, and will no longer attempt to insert them into the contents of script or style tags, or inside of any html tag. Also added a new favicon for the new release though you will need to delete, move, or rename the previous favicon and notice for it to work.