QoL additions for the new room functionality.
Room tabs will log in automatically if a cookie with a valid username and password is present, and room tags will open the logs of their respective room with the Open Logs player command. A Clean Logs admin command has also been added, which takes in a file size in KB. Any log files below that size are deleted. Use with discretion.