Added automatic cleanup for log files that only contain logins and Disconnect Other Instances button on login.
Any time a log file is replaced at the end of the day or the Shutdown command is run on the server, the log file of the previous day will be deleted if it only contained logons and logoffs. Also, the login screen now includes the Disconnect Other Instances button, which lets a user disconnect other client connections under their username to guarantee they can login.