Various QoL changes.
Interactive logs now alter the URL as the page changes so that refreshing still takes you to the same log. The space at the top of the character tab has been enlarged somewhat for easier right clicking. Up to the last ten IC posts now show up in the IC window upon login, along with a link to the log page of the day.