Added Safety Nets When Cannot Post
If you attempt to make a post when you cannot, i.e. you are muted or you are a Guest attempting to post to IC, the message box will remain open and the text will not be erased.
Also, you will now be informed when you are muted or unmuted.